requirement.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.3
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==1.0.4
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-reversion==3.0.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
gevent==1.3.7
greenlet==0.4.15
httplib2==0.12.0
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
lxml==4.2.6
Markdown==2.6.8
MarkupSafe==1.1.0
mysqlclient==1.3.14
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.4.0
PyJWT==1.7.1
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
